poc/backend/requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
lxml==6.0.1
requests==2.32.5
soupsieve==2.7
typing_extensions==4.15.0
urllib3==2.5.0
pinecone==7.3.0
pinecone-client==5.0.1
sentence-transformers==3.3.1
langchain-google-genai==2.0.9
langchain-huggingface==0.1.2
langchain-core==0.3.29
langchain_pinecone
langchain
python-dotenv==1.1.1
fastapi==0.115.2
uvicorn[standard]==0.33.0
pydantic==2.10.5
pydantic-settings==2.4.0